all the tests are now converted over to no longer depend on create jdbc-connection

+++ geode-connectors/src/acceptanceTest/java/org/apache/geode/connectors/jdbc/JdbcAsyncWriterIntegrationTest.java
@@ -81,11 +81,14 @@ public abstract class JdbcAsyncWriterIntegrationTest {
 
   private void closeDB() throws Exception {
     if (statement == null) {
-      statement = connection.createStatement();
+      if (connection != null) {
+        statement = connection.createStatement();
+      }
     }
-    statement.execute("Drop table " + REGION_TABLE_NAME);
-    statement.close();
-
+    if (statement != null) {
+      statement.execute("Drop table " + REGION_TABLE_NAME);
+      statement.close();
+    }
     if (connection != null) {
       connection.close();
     }
@@ -241,7 +244,8 @@ public abstract class JdbcAsyncWriterIntegrationTest {
   private SqlHandler createSqlHandler()
       throws ConnectionConfigExistsException, RegionMappingExistsException {
     return new SqlHandler(new TestableConnectionManager(), new TableMetaDataManager(),
-        TestConfigService.getTestConfigService(getConnectionUrl()));
+        TestConfigService.getTestConfigService(getConnectionUrl()),
+        new TestDataSourceFactory(getConnectionUrl()));
   }
 
 }
